refactor(logging): privatize full-context marker
