+++ src/logging/redact.test.ts
@@ -4,7 +4,6 @@ import os from "node:os";
 import path from "node:path";
 import { afterEach, describe, expect, it } from "vitest";
 import { withEnv } from "../test-utils/env.js";
-import { withFullContextToolPayloadRedaction } from "./redact-internal.js";
 import {
   getDefaultRedactPatterns,
   redactSecrets,
@@ -15,6 +14,7 @@ import {
   redactToolPayloadTextWithConfig,
   resolveRedactOptions,
 } from "./redact.js";
+import { withFullContextToolPayloadRedaction } from "./redact.test-support.js";
 import { registerSecretValueForRedaction } from "./secret-redaction-registry.js";
 import { resetSecretRedactionRegistryForTest } from "./secret-redaction-registry.test-support.js";
